split fetch out into manifest package
